State space more detailed
the 'state' header declares a hilberspace and a state struct, the
Hilbertspace has the low level definition of the C operators, the state
uses them to define the actual ones. The state also stores a reference
to a Hilbertspace. The hilbertspace needs to store partially the base of
the 2**nofOrbitals dimensional space, through the reference this is only
stored once.