Matrix Greensfunction computation (2 site DMFT only!)
I set up the "m-setup.hpp" header which defines an Setup struct
this creates and stores internally a 16x16 matrix (from given U and to
be computed V). In also stores the system state and (it can't yet) it
can evolve the system state by means of its Hamiltonian.

It can also act by C and Cd operators on the state. (not yet)