cleaning up: Makefiles
Instead of doing everything in one huge and confusing Makefile I'm
calling make in each subdirectory and split the compile and link stage.

The compile stage is a breeze: `g++ -MM a.cpp > .depend` fills .depend
with the make dependency list for a.o. We only need to include .depend
as a makefile into the makefile that compiles a.cpp, then a.cpp is
recompiled every time a header or inline is changed.

Linking however is worse, as we then need to know which *.o files
implement which header. This can not simply be done by reversing the
information in .depend, as this simply tells us which files need the
header, but not which actually implement it.

Solution: Makefiles in the main/ and test/ directories need to list
explicitly which files need to be linked into the *.prog/*.test files.
This sounds worse than it is: The rule

%.prog:
    g++ $^ -o $@

links everything that is in the list specified by "$^" which are the
dependencies of "$@" which is the target (the *.prog file). Now we only
need to give the correct list "$^" for each *.prog file like

main.prog: main.o ../src/qc.o ../src/util.o

then all the *.prog files are made the dependency of "all" and the usual
rule to compile "main.o" and similar is given:

%.o : %.cpp
    g++ -c $<