Adding adaptive stepsize
Freqency dependent greensfunction is dependant of the time resolution.
The adaptive timestep guarantees that the most prominent frequency is
given 200 support points and 50 periods of this frequency are sampled
(-> 10000 sample points are used).

This might be dangerous:
 * maybe a smaller timestep would reveal a more dominant high frequency
 * using much more than 50 periods of this frequency could show modulating
   low frequencies.

WARN: the quasiparticle weight is negative, but shouldn't be, as it is
  V**2 / t**2 (and both t and V are real).
  V is taken positive, is this correct?

Until now I was under the impression that there might be a bug with
computing the fft, this has been detected and cleared in the previous
commits.
 -> this bug is important