Huge change in QC::Rig
The implementation of the gates was somewhat ugly (I don't know if it's
now any different)

Was: privates void Rig::loop() and void Rig::ctrl() applied a void (*)(complex&,
complex&) while looping over the Rig and only differed in the check
function they used to decide wether the function should be applied
or not.
 -> now unified to public Rig& Rig::gate().
loop() ans ctrl() were called by the single and multi Qbit gates that
also were member function of the Rig class. -> bloat.
 -> moved the single Qbit gates (XYZH) to the QC namespace and gave it a
 type (unitary aka void (complex&, complex&) )
 -> when Rig::gate() is called with default args it applies a single Bit
 gate, when the trailing args are
    ...,index,ifset)
 the gate is applied if (ifset ? index==up : index==dw )

Qbit rotations: if Rig::gate is called with a double parameter after the
unitary parameter it rotates the qbit around the double paramater angle
(times PI) via
        cos(a) Id + i sin(a) P
when P is the pauli matrix provided.