making struct green store a pointer to the setup
instead of a reference. This enables to switch out the setup in order to
switch between calculation modes.

Will now rename the setup to calculate greensfunction with dense matrix
methods and base it on a generic

class Setup
{ public: double U,V; void set(/* */); Fourier::spec_t green(/* */) const; };

Will now move the convergence loop to struct green and rename it DMFT