Threat made real
the mainloop now lives in converge() which leaves the Twosite::setup in
a defined state (nullptr) such that it cant dangle if it is deleted
afterwards (only the calculation needs to have it set anyways).

It can also return the history of the parameter and ignores it if a
nullptr is passed for the storage.