Qmeth doesn't work nicely
minor changes; in search for bug with Qmeth::green() (result is
constant, indep of timestep size)

Current problems:
 * If we start at an arbitrary U,V config the step needs to be set
accordingly for green() to compute something that timestep() then can
examine and correct it nicely. We don't have good initial guesses which
leads to the routine break on occasion (the computation creates NaNs)

 * Qmeth::green() computes a constant value (-4i) for no apparent reason.
 * Qmeth::green() takes unnecessarily long
 * Twosite::ckparams() creates a negative V for the outcome of
 DenseED::green()
 * converge() reaches convergence at a seemingly arbitrary point for
 DenseED::green()

-> create storage of QC::Rig state to not compute |f>(t) over and over
again. (can cut down to 2/cycles of duration: N*(N-1)/2/N=(N-1)/2 )
-> make Twosite::timestep() create sensible new timesteps (if it changes
too hard, it's probably because the fourier trafo failed; want a minimum
of samples for one period)
-> find bugs in Qmeth::green() and DenseED::green() (haha...   ..oof.. )